WORKING: managed to fix the bugs. can now send and SMS upon detecting UWQ and maanage multiple alarms fairly easily. Can also send an SMS immediately, if back to safe but detetcs an unsafe WQ

+++ final_project/lib/States/src/States.h
@@ -120,7 +120,7 @@ class rmsClass {
       uint8_t get_totalStateChanges();
 
       void set_stateHistoryPercentage(RMSState stateOfInterest);
-      uint8_t get_stateHistoryPercentage(RMSState stateOfInterest);
+      float get_stateHistoryPercentage(RMSState stateOfInterest);
 
       void reset_History();
